Make sure Image post_save signal in process when Image is subclassed

+++ satchless/image/models.py
@@ -54,6 +54,7 @@ class Thumbnail(models.Model):
         unique_together = ('image', 'size')
 
 def original_changed(sender, instance, created, **kwargs):
-    instance.thumbnail_set.all().delete()
+    if isinstance(instance, Image):
+        instance.thumbnail_set.all().delete()
 
-models.signals.post_save.connect(original_changed, sender=Image)
+models.signals.post_save.connect(original_changed)
